Create pages for each endpoint
