+++ .env.development
@@ -1,2 +1,1 @@
-NETWORK_NAME=localhost
-GRAPHQL_ENDPOINT_URI=http://localhost:5000/graphql/
+GRAPHQL_ENDPOINTS=[{"name": "localhost", "uri": "http://localhost:5000/"}]
